test(j-s): Add appeal steps to search warrant tests
