+++ apps/system-e2e/src/tests/judicial-system/smoke/search-warrant-tests.spec.ts
@@ -5,6 +5,9 @@ import { test } from '../utils/judicialSystemTest'
 import { randomCourtCaseNumber, randomPoliceCaseNumber } from '../utils/helpers'
 import { urls } from '../../../support/urls'
 import { verifyRequestCompletion } from '../../../support/api-tools'
+import { appealCaseTest } from './shared/appeal-case'
+import { receiveAppealTest } from './shared/receive-appeal'
+import { coaJudgeCompletesCaseTest } from './shared/coa-complates-case'
 
 test.use({ baseURL: urls.judicialSystemBaseUrl })
 
@@ -92,58 +95,67 @@ test.describe.serial('Search warrant tests', () => {
     await expect(page).toHaveURL('/krofur')
   })
 
-  test.describe.serial('Search warrant court tests', () => {
-    test('court should receive search warrant request and make a ruling', async ({
-      judgePage,
-    }) => {
-      const page = judgePage
-      const yesterday = subDays(new Date(), 1).toLocaleDateString('is-IS', {
-        month: '2-digit',
-        day: '2-digit',
-        year: 'numeric',
-      })
-
-      await page.goto(`domur/rannsoknarheimild/mottaka/${caseId}`)
-      await page.getByTestId('courtCaseNumber').click()
-      await page.getByTestId('courtCaseNumber').fill(randomCourtCaseNumber())
-      await page
-        .getByText('Veldu dómara/aðstoðarmann *Veldu héraðsdómara')
-        .click()
-      await page.getByTestId('select-judge').getByText('Test Dómari').click()
-      await page.getByTestId('continueButton').click()
-      await expect(page).toHaveURL(`/domur/rannsoknarheimild/yfirlit/${caseId}`)
-      await page.getByTestId('continueButton').click()
-      await page.getByText('Fulltrúar málsaðila boðaðir').click()
-      await page.getByTestId('continueButton').click()
-      await page.getByTestId('modalPrimaryButton').click()
-      await page.getByText('Krafa samþykkt').click()
-      await page.getByTestId('continueButton').click()
-      await page.getByLabel('Dagsetning þingfestingar *').click()
-      await page.locator('input[id=courtStartDate]').fill(yesterday)
-      await page.getByTestId('courtStartDate-time').click()
-      await page.getByTestId('courtStartDate-time').fill('12:00')
-      await page
-        .locator('label')
-        .filter({ hasText: 'Varnaraðili unir úrskurðinum' })
-        .click()
-      await page
-        .locator('label')
-        .filter({ hasText: 'Sækjandi tekur sér lögboðinn frest' })
-        .click()
-      await page.getByLabel('Dagsetning uppkvaðningar *').click()
-      await page.locator('input[id=courtEndTime]').fill(yesterday)
-      await page
-        .getByLabel('Dagsetning uppkvaðningar *')
-        .click({ timeout: 1000 })
-      await page.getByTestId('courtEndTime-time').click()
-      await page.getByTestId('courtEndTime-time').fill('15:00')
-      await page.getByTestId('continueButton').click()
-      await expect(page).toHaveURL(
-        `/domur/rannsoknarheimild/stadfesta/${caseId}`,
-      )
-      await page.getByTestId('continueButton').click()
-      await page.getByTestId('modalPrimaryButton').click()
-      await expect(page).toHaveURL(`/krafa/yfirlit/${caseId}`)
+  test('court should receive search warrant request and make a ruling', async ({
+    judgePage,
+  }) => {
+    const page = judgePage
+    const yesterday = subDays(new Date(), 1).toLocaleDateString('is-IS', {
+      month: '2-digit',
+      day: '2-digit',
+      year: 'numeric',
     })
+
+    await page.goto(`domur/rannsoknarheimild/mottaka/${caseId}`)
+    await page.getByTestId('courtCaseNumber').click()
+    await page.getByTestId('courtCaseNumber').fill(randomCourtCaseNumber())
+    await page
+      .getByText('Veldu dómara/aðstoðarmann *Veldu héraðsdómara')
+      .click()
+    await page.getByTestId('select-judge').getByText('Test Dómari').click()
+    await page.getByTestId('continueButton').click()
+    await expect(page).toHaveURL(`/domur/rannsoknarheimild/yfirlit/${caseId}`)
+    await page.getByTestId('continueButton').click()
+    await page.getByText('Fulltrúar málsaðila boðaðir').click()
+    await page.getByTestId('continueButton').click()
+    await page.getByTestId('modalPrimaryButton').click()
+    await page.getByText('Krafa samþykkt').click()
+    await page.getByTestId('continueButton').click()
+    await page.getByLabel('Dagsetning þingfestingar *').click()
+    await page.locator('input[id=courtStartDate]').fill(yesterday)
+    await page.getByTestId('courtStartDate-time').click()
+    await page.getByTestId('courtStartDate-time').fill('12:00')
+    await page
+      .locator('label')
+      .filter({ hasText: 'Varnaraðili unir úrskurðinum' })
+      .click()
+    await page
+      .locator('label')
+      .filter({ hasText: 'Sækjandi tekur sér lögboðinn frest' })
+      .click()
+    await page.getByLabel('Dagsetning uppkvaðningar *').click()
+    await page.locator('input[id=courtEndTime]').fill(yesterday)
+    await page.getByLabel('Dagsetning uppkvaðningar *').click({ timeout: 1000 })
+    await page.getByTestId('courtEndTime-time').click()
+    await page.getByTestId('courtEndTime-time').fill('15:00')
+    await page.getByTestId('continueButton').click()
+    await expect(page).toHaveURL(`/domur/rannsoknarheimild/stadfesta/${caseId}`)
+    await Promise.all([
+      verifyRequestCompletion(page, '/api/graphql', 'RequestRulingSignature'),
+      page.getByTestId('continueButton').click(),
+    ])
+  })
+
+  test('prosecutor should appeal case', async ({ prosecutorPage }) => {
+    await appealCaseTest(prosecutorPage, caseId)
+  })
+
+  test('judge should receive appealed case', async ({ judgePage }) => {
+    await receiveAppealTest(judgePage, caseId)
+  })
+
+  test('coa judge should submit decision in search warrant appeal case', async ({
+    coaPage,
+  }) => {
+    await coaJudgeCompletesCaseTest(coaPage, caseId)
   })
 })
