+++ apps/system-e2e/src/tests/judicial-system/smoke/shared/receive-appeal.ts
@@ -1,0 +1,11 @@
+import { Page } from '@playwright/test'
+
+export async function receiveAppealTest(page: Page, caseId: string) {
+  await page.goto(`krafa/yfirlit/${caseId}`)
+  await page
+    .getByRole('button', {
+      name: 'Senda tilkynningu um kæru til Landsréttar',
+    })
+    .click()
+  await page.getByTestId('modalPrimaryButton').click()
+}
