+++ apps/system-e2e/src/tests/judicial-system/utils/helpers.ts
@@ -1,6 +1,3 @@
-// import PDFDocument from 'pdfkit'
-// import getStream from 'get-stream'
-
 export function randomPoliceCaseNumber() {
   return `007-${new Date().getFullYear()}-${Math.floor(Math.random() * 100000)}`
 }
@@ -12,13 +9,6 @@ export function randomCourtCaseNumber() {
 export function randomAppealCaseNumber() {
   return `${Math.floor(Math.random() * 1000)}/${new Date().getFullYear()}`
 }
-
-// export async function createPdf(text: string) {
-//   const doc = new PDFDocument()
-//   doc.text(text)
-//   doc.end()
-//   return await getStream.buffer(doc)
-// }
 
 export async function createFakePdf(title: string) {
   return {
